Выносит отрисовку дней в отдельный модуль
